zebra: revert pimd nexthop_active change
this change seems bogus to me; kernel routes are routes added by other
routing daemons. the intention seems to be to make recursive nexthops
on kernel routes work, but either this needs to apply to static routes
as well or it's simply wrong.

either way this needs to be inspected outside of the pimd merge.

 * zebra/zebra_rib.c: reverted